+++ business/core/post/post.go
@@ -19,7 +19,7 @@ var (
 type Storer interface {
 	Create(ctx context.Context, post Post) (sql.Result, error)
 	Delete(ctx context.Context, post Post) error
-	QueryById(ctx context.Context, id uint64) (Post, error)
+	QueryById(ctx context.Context, id int64) (Post, error)
 	Query(ctx context.Context, filter QueryFilter, orderBy order.OrderBy, pageNumber int, rowsPerPage int) ([]Post, error)
 }
 
@@ -67,11 +67,11 @@ func (c *Core) Create(ctx context.Context, newPost NewPost) (Post, error) {
 	}
 
 	post := Post{
-		Id: id,
+		Id: int64(id),
 		UserId: newPost.UserId,
 		Title: newPost.Title,
 		Description: newPost.Description,
-		ContentId: contentId,
+		ContentId: int64(contentId),
 		Content: newPost.Content,
 		CreatedAt: now,
 		UpdatedAt: now,
@@ -96,7 +96,7 @@ func (c *Core) Delete(ctx context.Context, post Post) error {
 	return nil
 }
 
-func (c *Core) QueryById(ctx context.Context, id uint64) (Post, error) {
+func (c *Core) QueryById(ctx context.Context, id int64) (Post, error) {
 	post, err := c.storer.QueryById(ctx, id)
 	if err != nil {
 		return Post{}, err
